Add authorization check to delete and update actions(forgot to add this in the last commit) comparing the logged in user with the creator of the post

+++ controllers/feed.js
@@ -157,6 +157,11 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       // check logged in User
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
